pyproject.toml
[tool.poetry]
name = "cloudships"
version = "0.1.0"
description = ""
authors = ["Alec McGavin <dev@mcgav.in>"]

[tool.poetry.dependencies]
python = "^3.8"
waitress = "^1.4.3"
google-cloud-secret-manager = "^0.2.0"
google-cloud-datastore = "^1.11.1"
django = "==3.0.4"
django-environ = "^0.4.5"
psycopg2-binary = "^2.8.5"
django-storages = {extras = ["google"], version = "^1.9.1"}
djangorestframework = "^3.11.0"
django-extensions = "^2.2.9"
whitenoise = "^5.0.1"

[tool.poetry.dev-dependencies]
ipython = "^7.13.0"

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
